pyproject.toml
[project]
name = "smart-travel-planner"
version = "0.1.0"
description = "AI-powered travel planner with ML classification, RAG retrieval, and LangGraph agent"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "sqlalchemy[asyncio]>=2.0.35",
    "greenlet>=3.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pgvector>=0.2.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.1.0",
    "python-dateutil>=2.8.2",
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langsmith>=0.1.0",
    "openai>=1.50.0",
    "anthropic>=0.30.0",
    "scikit-learn>=1.5.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "xgboost>=2.1.0",
    "imbalanced-learn>=0.12.0",
    "joblib>=1.4.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.3",
    "tenacity>=9.0.0",
    "sentence-transformers>=3.0.1",
    "tqdm>=4.66.5",
    "cachetools>=5.5.0",
    "aiocache>=0.12.0",
    "pydantic>=2.9.0",
    "email-validator>=2.0.0",
    "pydantic-settings>=2.4.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.4.0",
    "click>=8.1.0",
    "tiktoken>=0.7.0",
    "matplotlib>=3.10.9",
    "seaborn>=0.13.2",
    "psycopg2-binary>=2.9.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
    "black>=24.8.0",
    "mypy>=1.11.0",
    "pre-commit>=3.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.23.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["backend/tests"]
addopts = "-v --cov=backend/app --cov-report=term-missing"
